requirements.txt
uagents>=0.22.3
python-dotenv
requests
pyjwt
